requirements.txt
asgiref==3.5.2
Django==4.0.5
django-filter==22.1
djangorestframework==3.13.1
factory-boy==3.2.1
psycopg2-binary==2.9.3